Add QMetaType::QVariant as QVariant type to COM interface
It should be possible to explicitly return a VARIANT in the COM
interface. This is required when you want the calling program not to
change the output data type.

To return a (64 bit) pointer you want to use a uint64 datatype. However,
an uint64 in a VARIANT is converted by for instance MATLAB, to a double
data type which is not desirable for a pointer.
When returning a VARIANT, which has a VARIANT with uint64 data MATLAB
will treat it as an actual uint64.

This used to work without this patch in Qt 4.7.4, but not anymore in Qt
5.1

Change-Id: I847341ac62ae8608cd56dac79d2e732087340a6e