Search for platform plugins near a COM object library location
When using a COM object library, you want to have the platform folder
located next to the COM dll file, just like the other Qt*dll dependencies.
However for an in-process COM object, this location is not taken in to
account. Also you do not want to bother the user with the
QT_QPA_PLATFORM_PLUGIN_PATH environment variable.

Therefore when creating a COM server the current dll path is added to the
librarypath.

Change-Id: I80de155aa00d719e2c78f29b79a913e342737aa2